김기현 - groupchat 파트 group 패키지로 병합 완료
