+++ src/main/java/com/sist/web/group/mapper/GroupMapper.java
@@ -1,8 +1,13 @@
 package com.sist.web.group.mapper;
 
 import java.util.List;
+
+import org.apache.ibatis.annotations.Insert;
+import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
+import org.apache.ibatis.annotations.SelectKey;
 import com.sist.web.group.dto.GroupDTO;
+import com.sist.web.groupchat.dto.GroupMemberDTO;
 
 public interface GroupMapper {
 		//그룹
@@ -16,11 +21,22 @@ public interface GroupMapper {
 		public GroupDTO selectGroupDetail(int group_no);
 		
 		// 동적 쿼리 작성 예정
-		public void groupInsertData(GroupDTO vo);
+		public void groupInsertData(GroupDTO dto);
 		
 		@Select("SELECT p_group_no_seq.currval FROM DUAL")
 		public int groupCurrentNodata();
 		
+		@Insert("INSERT INTO p_group (group_no, group_name, description, owner) "
+				+ "VALUES(p_group_no_seq.nextval, #{group_name}, #{description}, #{owner})")
+		@SelectKey(statement = "SELECT p_group_no_seq.currval FROM dual", keyProperty = "group_no", before = false, resultType = Integer.class)
+		public void insertGroup(GroupDTO dto);
 		
+		public void insertGroupMember(GroupMemberDTO dto);
+		
+		@Select("SELECT g.group_no, group_name FROM p_group g JOIN p_group_member m ON g.group_no = m.group_no WHERE m.user_no = #{user_no}")
+		public List<GroupDTO> selectGroup(String userNo);
+		
+		@Select("SELECT group_no, user_no, nickname FROM p_group_member WHERE group_no = #{group_no}")
+		public List<GroupMemberDTO> selectGroupMemberAllByGroupNo(@Param("group_no") int groupNo);
 		
 }
