+++ src/main/java/com/sist/web/groupchat/mapper/GroupChatMapper.java
@@ -13,7 +13,7 @@ public interface GroupChatMapper {
 	@Insert("INSERT INTO p_group_msg (message_no, group_no, sender_no, content, sender_nickname) "
 		  + "VALUES (p_gm_mno_seq.nextval, #{group_no}, #{sender_no}, #{content}, #{sender_nickname})")
 	@SelectKey(statement = "SELECT p_gm_mno_seq.currval FROM dual", keyProperty = "message_no", before = false, resultType = Long.class)
-	public void insertGroupChatMessage(GroupChatDTO vo);
+	public void insertGroupChatMessage(GroupChatDTO dto);
 	
 	public List<GroupChatDTO> selectLatestMessageByGroupNo(@Param("group_no") int groupNo, @Param("last_message_no") Long lastMessageNo);
 	
@@ -21,17 +21,12 @@ public interface GroupChatMapper {
 	public GroupChatDTO selectGroupChatByMessageNo(@Param("message_no") long messageNo);
 	
 	
-	public void insertGroupMember(GroupMemberDTO vo);
 	
-	@Select("SELECT g.group_no, group_name FROM p_group g JOIN p_group_member m ON g.group_no = m.group_no WHERE m.user_no = #{user_no}")
-	public List<GroupDTO> selectGroup(String userNo);
 	
-	@Select("SELECT group_no, user_no, nickname FROM p_group_member WHERE group_no = #{group_no}")
-	public List<GroupMemberDTO> selectGroupMemberAllByGroupNo(@Param("group_no") int groupNo);
 	
-	@Insert("INSERT INTO p_group (group_no, group_name, description, owner) "
-			+ "VALUES(p_group_no_seq.nextval, #{group_name}, #{description}, #{owner})")
-	@SelectKey(statement = "SELECT p_group_no_seq.currval FROM dual", keyProperty = "group_no", before = false, resultType = Integer.class)
-	public void insertGroup(GroupDTO vo);
+	
+	
+	
+	
 	
 }
