+++ src/main/java/com/sist/web/groupchat/service/GroupChatService.java
@@ -9,8 +9,7 @@ import com.sist.web.groupchat.dto.GroupMemberDTO;
 public interface GroupChatService {
 	void saveAndSendGroupChatMessage(GroupChatDTO vo);
 	List<GroupChatDTO> getLatestMessageByGroupNo(int groupNo, Long lastMessageNo);
-	void createGroup(GroupDTO vo);
-	void addGroupMember(GroupMemberDTO vo);
-	List<GroupDTO> getGroupAll(String userId);
-	List<GroupMemberDTO> getGroupMemberAllByGroupNo(int groupNo);
+	
+	
+	
 }
